// netlify/functions/bookings/bookings-schema.d.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
  "/bookings": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Retrieve all bookings for the given duration
     * @description Returns all bookings within a given date range, or four weeks starting from this week
     */
    get: {
      parameters: {
        query?: {
          start?: string;
          end?: string;
        };
        header?: never;
        path?: never;
        cookie?: never;
      };
      requestBody?: never;
      responses: {
        /** @description A list of bookings */
        200: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            "application/json": {
              /** @example Bosse Bredsladd */
              name: string;
              /** @example [
               *       "Carpenter",
               *       "Painter"
               *     ] */
              professions: ("Carpenter" | "Electrician" | "Painter" | "Mason" | "Plumber")[];
              bookings: {
                /**
                 * @example Painter
                 * @enum {string}
                 */
                activity: "Carpenter" | "Electrician" | "Painter" | "Mason" | "Plumber" | "Other";
                /** Format: date */
                from: string;
                /** Format: date */
                to: string;
                /**
                 * @example 100
                 * @enum {integer}
                 */
                percentage: 50 | 100;
                /**
                 * @example Booked
                 * @enum {string}
                 */
                status: "Absent" | "Booked" | "Preliminary";
              }[];
            }[];
          };
        };
      };
    };
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: never;
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}
export type $defs = Record<string, never>;
export type operations = Record<string, never>;